+++ hutool-core/src/main/java/cn/hutool/core/exceptions/ExceptionUtil.java
@@ -18,10 +18,11 @@ import cn.hutool.core.util.StrUtil;
  *
  */
 public final class ExceptionUtil {
-	
+
 	private static final String NULL = "null";
-	
-	private ExceptionUtil(){};
+
+	private ExceptionUtil() {
+	};
 
 	/**
 	 * 获得完整消息，包括异常名
@@ -30,13 +31,12 @@ public final class ExceptionUtil {
 	 * @return 完整消息
 	 */
 	public static String getMessage(Throwable e) {
-		if(null == e) {
+		if (null == e) {
 			return NULL;
 		}
 		return StrUtil.format("{}: {}", e.getClass().getSimpleName(), e.getMessage());
 	}
-	
-	
+
 	/**
 	 * 获得完整消息，包括异常名
 	 * 
@@ -46,22 +46,26 @@ public final class ExceptionUtil {
 	public static String getSimpleMessage(Throwable e) {
 		return (null == e) ? NULL : e.getMessage();
 	}
-	
+
 	/**
 	 * 使用运行时异常包装编译异常
+	 * 
 	 * @param throwable 异常
 	 * @return 运行时异常
 	 */
-	public static RuntimeException wrapRuntime(Throwable throwable){
-		if(throwable instanceof RuntimeException){
+	public static RuntimeException wrapRuntime(Throwable throwable) {
+		if (throwable instanceof RuntimeException) {
 			return (RuntimeException) throwable;
-		}else{
-			return new RuntimeException(throwable);
 		}
+		if (throwable instanceof Error) {
+			throw (Error) throwable;
+		}
+		return new RuntimeException(throwable);
 	}
-	
+
 	/**
 	 * 包装一个异常
+	 * 
 	 * @param throwable 异常
 	 * @param wrapThrowable 包装后的异常类
 	 * @return 包装后的异常
@@ -69,7 +73,7 @@ public final class ExceptionUtil {
 	 */
 	@SuppressWarnings("unchecked")
 	public static <T extends Throwable> T wrap(Throwable throwable, Class<T> wrapThrowable) {
-		if(wrapThrowable.isInstance(throwable)) {
+		if (wrapThrowable.isInstance(throwable)) {
 			return (T) throwable;
 		}
 		return ReflectUtil.newInstance(wrapThrowable, throwable);
@@ -77,6 +81,7 @@ public final class ExceptionUtil {
 
 	/**
 	 * 剥离反射引发的InvocationTargetException、UndeclaredThrowableException中间异常，返回业务本身的异常
+	 * 
 	 * @param wrapped 包装的异常
 	 * @return 剥离后的异常
 	 */
@@ -92,17 +97,17 @@ public final class ExceptionUtil {
 			}
 		}
 	}
-	
+
 	/**
 	 * 获取当前栈信息
 	 * 
 	 * @return 当前栈信息
 	 */
 	public static StackTraceElement[] getStackElements() {
-//		return (new Throwable()).getStackTrace();
+		// return (new Throwable()).getStackTrace();
 		return Thread.currentThread().getStackTrace();
 	}
-	
+
 	/**
 	 * 获取指定层的堆栈信息
 	 * 
@@ -112,7 +117,7 @@ public final class ExceptionUtil {
 	public static StackTraceElement getStackElement(int i) {
 		return getStackElements()[i];
 	}
-	
+
 	/**
 	 * 获取入口堆栈信息
 	 * 
@@ -123,82 +128,122 @@ public final class ExceptionUtil {
 		final StackTraceElement[] stackElements = getStackElements();
 		return stackElements[stackElements.length - 1];
 	}
-	
+
 	/**
 	 * 堆栈转为单行完整字符串
+	 * 
 	 * @param throwable 异常对象
 	 * @return 堆栈转为的字符串
 	 */
-	public static String stacktraceToOneLineString(Throwable throwable){
+	public static String stacktraceToOneLineString(Throwable throwable) {
 		return stacktraceToOneLineString(throwable, 3000);
 	}
-	
+
 	/**
 	 * 堆栈转为单行完整字符串
+	 * 
 	 * @param throwable 异常对象
 	 * @param limit 限制最大长度
 	 * @return 堆栈转为的字符串
 	 */
-	public static String stacktraceToOneLineString(Throwable throwable, int limit){
+	public static String stacktraceToOneLineString(Throwable throwable, int limit) {
 		Map<Character, String> replaceCharToStrMap = new HashMap<>();
 		replaceCharToStrMap.put(StrUtil.C_CR, StrUtil.SPACE);
 		replaceCharToStrMap.put(StrUtil.C_LF, StrUtil.SPACE);
 		replaceCharToStrMap.put(StrUtil.C_TAB, StrUtil.SPACE);
-		
+
 		return stacktraceToString(throwable, limit, replaceCharToStrMap);
 	}
-	
+
 	/**
 	 * 堆栈转为完整字符串
+	 * 
 	 * @param throwable 异常对象
 	 * @return 堆栈转为的字符串
 	 */
-	public static String stacktraceToString(Throwable throwable){
+	public static String stacktraceToString(Throwable throwable) {
 		return stacktraceToString(throwable, 3000);
 	}
-	
+
 	/**
 	 * 堆栈转为完整字符串
+	 * 
 	 * @param throwable 异常对象
 	 * @param limit 限制最大长度
 	 * @return 堆栈转为的字符串
 	 */
-	public static String stacktraceToString(Throwable throwable, int limit){
+	public static String stacktraceToString(Throwable throwable, int limit) {
 		return stacktraceToString(throwable, limit, null);
 	}
-	
+
 	/**
 	 * 堆栈转为完整字符串
+	 * 
 	 * @param throwable 异常对象
 	 * @param limit 限制最大长度
 	 * @param replaceCharToStrMap 替换字符为指定字符串
 	 * @return 堆栈转为的字符串
 	 */
-	public static String stacktraceToString(Throwable throwable, int limit, Map<Character, String> replaceCharToStrMap){
+	public static String stacktraceToString(Throwable throwable, int limit, Map<Character, String> replaceCharToStrMap) {
 		final FastByteArrayOutputStream baos = new FastByteArrayOutputStream();
 		throwable.printStackTrace(new PrintStream(baos));
 		String exceptionStr = baos.toString();
 		int length = exceptionStr.length();
-		if(limit > 0 && limit < length){
+		if (limit > 0 && limit < length) {
 			length = limit;
 		}
-		
-		if(CollectionUtil.isNotEmpty(replaceCharToStrMap)){
+
+		if (CollectionUtil.isNotEmpty(replaceCharToStrMap)) {
 			final StringBuilder sb = StrUtil.builder();
 			char c;
 			String value;
-			for(int i = 0; i < length; i++){
+			for (int i = 0; i < length; i++) {
 				c = exceptionStr.charAt(i);
 				value = replaceCharToStrMap.get(c);
-				if(null != value){
+				if (null != value) {
 					sb.append(value);
-				}else{
+				} else {
 					sb.append(c);
 				}
 			}
 			return sb.toString();
-		}else{
+		} else {
 			return exceptionStr;
 		}
+	}
+
+	/**
+	 * 判断是否由指定异常类引起
+	 * 
+	 * @param throwable 异常
+	 * @param causeClasses 定义的引起异常的类
+	 * @return 是否由指定异常类引起
+	 * @since 4.1.13
+	 */
+	@SuppressWarnings("unchecked")
+	public static boolean isCausedBy(Throwable throwable, Class<? extends Exception>... causeClasses) {
+		return null != getCausedBy(throwable, causeClasses);
+	}
+	
+	/**
+	 * 获取由指定异常类引起的异常
+	 * 
+	 * @param throwable 异常
+	 * @param causeClasses 定义的引起异常的类
+	 * @return 是否由指定异常类引起
+	 * @since 4.1.13
+	 */
+	@SuppressWarnings("unchecked")
+	public static Throwable getCausedBy(Throwable throwable, Class<? extends Exception>... causeClasses) {
+		Throwable cause = throwable;
+		while (cause != null) {
+			for (Class<? extends Exception> causeClass : causeClasses) {
+				if (causeClass.isInstance(cause)) {
+					return cause;
+				}
+			}
+			cause = cause.getCause();
+		}
+		return null;
 	}
 }