+++ hutool-core/src/main/java/cn/hutool/core/thread/ExecutorBuilder.java
@@ -110,10 +110,13 @@ public class ExecutorBuilder implements Builder<ThreadPoolExecutor> {
 	}
 
 	/**
-	 * 设置当线程阻塞（block）时的处理器
+	 * 设置当线程阻塞（block）时的异常处理器，所谓线程阻塞既线程池和等待队列已满，无法处理线程时采取的策略
+	 * <p>
+	 * 此处可以使用JDK预定义的几种策略，见{@link RejectPolicy}枚举
 	 * 
 	 * @param handler {@link RejectedExecutionHandler}
 	 * @return this
+	 * @see RejectPolicy
 	 */
 	public ExecutorBuilder setHandler(RejectedExecutionHandler handler) {
 		this.handler = handler;